allow floats as bar values

+++ csrc/bar_chart.h
@@ -10,10 +10,10 @@
 typedef struct BarChart Bar;
 
 /* FUNCTION PROTOTYPES */
-int get_number (Bar *series);
+float get_number (Bar *series);
 char *get_name (Bar *series);
-void set_number (Bar *series, int number);
+void set_number (Bar *series, float number);
 void set_name (Bar *series, char *name);
 void print_bar(Bar *series, char *chart, int x_offset, float scale);
-Bar *new_bar (int number, char *name);
+Bar *new_bar (float number, char *name);
 #endif
